refactor: move all commands to using defcom
A bit annoying to have to import this everywhere, and we now need to
make sure everything is required somewhere (ralphie.core for now).

Otherwise, this lowers the cost to adding new commands, which is a win.

It may also allow for smaller/filtered commands by removing some
requires, which might speed things up?