feat: build-and-install-uberscript
Between the existing executable, a built uberjar, and a built
uberscript, the uberscript appears to be the most performant. This adds
a command for building and installing it, and updates the existing
install command to build to ralphie-dev.

Gives a nice speed boost to all existing ralphie commands, and prevents
bugs in those commands when hacking on new features (until the are
'released' via this new command again.)