feat: readme rewrite, remove crazy shell-bang preamble
For now, deletes the split/auto-gened readme - it was getting in the way of
just-writing-the-docs. I'd rather auto-gen the docs as org files via
org-crud - seems much more managable at this point.

The shell intro on ralphie.core makes that file executable, but at this
point it's easy enough to just use `bb` directly - the file itself doesn't
need that power. The readme was updated to including working exec
commands, and it can be wrapped in another script if it's really
necessary to support a dev-bin that accessible.