refactor: rofi ns refactored to use bb/process
Interestingly i crashed my os several times while working on this.
Not sure if it was rofi waiting for input or bb/process waiting to
consume from rofi. had to switch to a tty and reboot lightdm.

This effectively wraps the rofi api, so hopefully we don't have to think
about it ever again.