wip: tried to get a carve/microscript-install going
But did not succeed. struggling to get a ($ carve) to error/succeed in
the repl. Installed the latest jdk11-graalvm-bin at one point. Need to
learn more about carve, and whether it's the right tool for this.

Cleans up a few cli/command apis, and moves deps from local roots to
mvn/github.