feat: install-micro command for creating minimal uberscripts
Supports creating and installing minimal uberscripts for any command.
A temporary .clj file is created with a -main function that calls the
command's handler with the passed args. That ns is used to create an
uberscript, which is then carved down to only the used functions.
The final step installs the script on the path, at `ralphie-<cmd-name>`.

I tweaked defcom to support a bit more code to do this, which resulted
in a much larger filesize for toggle-scratchpad. The speed is still
quite fast, but not as fast as it was - more digging to do there.

Also cleans up the core.-main func and impls a draft of a rofi-dev
command. Not sure why it's not working.