fix: rofi nil-punning on nothing selected, handling linebreaks
Updates the rofi func to prevent a throw when nothing is selected. For
now this masks other errors that could come from rofi itself...

Also switches to a non-newline '\n' seperator to prevent unexpected
overwriting, and allow for decent doc strings to show up. Rofi doesn't
appear to support dynamic per-row height, which is too bad.